pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "galaxy-morphology-classification"
version = "0.1.0"
description = "Professional repository for galaxy morphology classification."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Daniel", email = "proyectostech3@gmail.com" }
]
license = { text = "MIT" }
dependencies = [
  "fastapi",
  "kagglehub",
  "matplotlib",
  "numpy",
  "pandas",
  "pillow",
  "pydantic",
  "python-multipart",
  "requests",
  "scikit-learn",
  "tensorflow",
  "uvicorn",
]

[tool.setuptools]
include-package-data = true